moved question properties
