+++ eq-author/src/App/page/Design/QuestionPageEditor/QuestionProperties/HelpModal.js
